requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
lxml==5.3.0
pillow==10.4.0
requests==2.32.3
